requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-google-genai
langchain-text-splitters
cohere
ragas
datasets
rapidfuzz
tqdm
pypdf
faiss-cpu
python-dotenv
streamlit
openai
langchain-classic
langchain-cohere
rank_bm25